treewide: update licenses
A linter is now complaining that Apache 2.0 is not valid for
some of our C files, so move the license lines to currently
acceptable licenses.

The tester main.c file is copy/pasted with an Intel copyright, so we
can't change the license and have to add it to the allowlist. (It's
small enough that maybe it should be rewritten with our own copyright
and license at some point, though.)

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>